refactor: extract dnd logic
